start on course, get spec barely passing
