+++ .gitignore
@@ -1,4 +1,5 @@
 *.env
+yarn-error.log
 node_modules
 bower_components
 dist
