+++ package.json
@@ -61,7 +61,9 @@
     "@sentry/browser": "^6.0.3",
     "@types/classnames": "^2.2.11",
     "@types/jest": "^26.0.20",
+    "@types/loadjs": "^4.0.0",
     "@types/lodash": "^4.14.168",
+    "@types/luxon": "^1.26.2",
     "@types/node-fetch": "^2.5.8",
     "@types/qs": "^6.9.6",
     "@types/react-redux": "^7.1.16",
@@ -97,12 +99,14 @@
     "jest-axe": "^3.2.0",
     "jest-changed-files": "^24.9.0",
     "jest-cli": "24.9.0",
+    "jest-fetch-mock": "^3.0.3",
     "jest-playwright-preset": "^1.4.2",
     "keymaster": "1.6.2",
     "lazy-get-decorator": "^2.2.0",
     "loadjs": "3.6.1",
     "lodash": "^4.17.19",
     "loglevel": "1.6.3",
+    "luxon": "^1.26.0",
     "markdown-it": "9.1.0",
     "markdown-it-regexp": "0.4.0",
     "mime-types": "2.1.24",
@@ -111,7 +115,7 @@
     "mobx-react": "^7.1.0",
     "mobx-utils": "^6.0.3",
     "mockdate": "^2.0.5",
-    "modeled-mobx": "0.0.3-0",
+    "modeled-mobx": "0.0.3-2",
     "moment": "2.24.0",
     "moment-range": "^4.0.2",
     "moment-timezone": "0.5.23",
