+++ shared/src/api/index.ts
@@ -1,12 +1,7 @@
-import { first, template } from 'lodash'
+import interpolate from 'interpolate'
 
-type HttpMethod = 'GET' | 'POST' | 'PUT' | 'DELETE'
-type RequestOptions = { params?: any, data?: any }
-type RequestReducerFunc<T> = (obj: T) => RequestOptions // eslint-disable-line no-unused-vars
-
-type Collection<T> = {
-    items: T[]
-}
+export type HttpMethod = 'GET' | 'POST' | 'PUT' | 'DELETE'
+export type RequestOptions = { params?: any, data?: any }
 
 class ApiError extends Error {
     request: string
@@ -25,12 +20,23 @@ const baseUrl = process.env.BACKEND_SERVER_URL ?
     process.env.BACKEND_SERVER_URL : window.location.port === '8000' ?
         'http://localhost:3001/api' : `${window.location.origin}/api`;
 
-export const request = async <RetT>(method: HttpMethod, urlPattern: string, opts?: RequestOptions): Promise<RetT> => {
-    let req: { method: HttpMethod, body?: any } = { method }
-    if (opts?.data) {
+const HttpMethodMatcher = /^(GET|POST|PATCH|PUT|DELETE)\s+/
+
+export const request = async<RetT>(
+    urlPattern: string,
+    opts: RequestOptions = {}
+): Promise<RetT> => {
+    let method = 'GET'
+    const methodMatch = urlPattern.match(HttpMethodMatcher)
+    if (methodMatch) {
+        method = methodMatch[1]
+        urlPattern = urlPattern.replace(HttpMethodMatcher, '')
+    }
+    let req: { method: string, body?: any } = { method }
+    if (opts.data) {
         req.body = JSON.stringify(opts.data)
     }
-    const url = template(urlPattern)(opts?.params || {})
+    const url = interpolate(urlPattern, opts.params)
     const resp = await fetch(`${baseUrl}/${url}`, req)
     const respJson = await resp.json()
     if (resp.ok) {
@@ -39,4 +45,3 @@ export const request = async <RetT>(method: HttpMethod, urlPattern: string, opts
 
     throw new ApiError(`${method} ${url}`, resp, opts)
 }
-
