+++ shared/src/model/exercise/attachment.ts
@@ -1,0 +1,16 @@
+import type { Exercise } from '../course'
+import {
+    BaseModel, identifier, field, belongsTo,
+} from '../../model';
+
+
+export default
+    class ExerciseAttachment extends BaseModel {
+
+    @model(Exercise) exercise;
+
+    @field asset;
+
+
+
+}
