+++ tutor/specs/courses-test-data.js
@@ -1,5 +1,6 @@
 import CoursesMap from '../src/models/courses-map';
 import moment from 'moment';
+import { runInAction } from 'mobx';
 
 const STUDENT_COURSE_ONE_MODEL = {
     id: '1',
@@ -176,6 +177,6 @@ const MASTER_COURSES_LIST = [
 const TUTOR_HELP = 'http://openstax.force.com/support?l=en_US&c=Products%3ATutor';
 const CONCEPT_COACH_HELP = 'http://openstax.force.com/support?l=en_US&c=Products%3AConcept_Coach';
 
-const bootstrapCoursesList = () => CoursesMap.bootstrap(MASTER_COURSES_LIST);
+const bootstrapCoursesList = () => runInAction(() => CoursesMap.bootstrap(MASTER_COURSES_LIST));
 
 export { STUDENT_COURSE_ONE_MODEL, TEACHER_COURSE_TWO_MODEL, STUDENT_ARCHIVED_COURSE, TEACHER_AND_STUDENT_COURSE_THREE_MODEL, MASTER_COURSES_LIST, TUTOR_HELP, CONCEPT_COACH_HELP, TEACHER_PAST_COURSE, STUDENT_PAST_COURSE, bootstrapCoursesList };
