+++ tutor/src/models/definitions.d.ts
@@ -1,0 +1,2 @@
+/// <reference path="course.ts" />
+/// <reference path="./course/period.ts" />
